:sparkles: Players can now pick up and drop items
We use our raycast to locate the nearest object to the player and we then use the interact key to pick it up, we make sure it stays a certain distance in front of the player at all times.

Once we press the interact key again, we drop the item.